requirements.txt
Django
djangorestframework
django-extensions
flake8
mock
psycopg2
requests
uwsgi
